agents/launch_report/launch_report/requirements.txt
requests
playwright
Pillow
python-dotenv
